fix color scheme and added sign out button to settings menu
