+++ app/src/main/java/com/example/team3_1/DriverHomePage.java
@@ -3,11 +3,15 @@ package com.example.team3_1;
 import androidx.annotation.NonNull;
 import androidx.annotation.Nullable;
 import androidx.appcompat.app.AppCompatActivity;
+import androidx.appcompat.widget.Toolbar;
 
 import android.app.Activity;
 import android.content.Intent;
 import android.os.Bundle;
 import android.util.Log;
+import android.view.Menu;
+import android.view.MenuInflater;
+import android.view.MenuItem;
 import android.widget.Button;
 
 import com.okta.oidc.AuthenticationPayload;
@@ -22,32 +26,36 @@ public class DriverHomePage extends AppCompatActivity {
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
         setContentView(R.layout.activity_driver_home_page);
+        Toolbar toolbar = (Toolbar) findViewById(R.id.include);
+        toolbar.setTitle("Driver");
+        setSupportActionBar(toolbar);
 
-        Button signOutBtn = findViewById(R.id.signOutBtn);
-        Activity thisclass = this;
-        signOutBtn.setOnClickListener(v -> {
-            //MainActivity.oktaManager.signOut(thisclass);
-            MainActivity.oktaManager.logOut(this, signOutCallback(thisclass));
-
-        });
     }
 
-    private RequestCallback<Integer, AuthorizationException> signOutCallback(Activity activity) {
-        RequestCallback object = new RequestCallback() {
-            @Override
-            public void onSuccess(@NonNull Object result) {
-                Intent intent = new Intent(activity, LoginActivity.class);
-                startActivity(intent);
-                finish();
-            }
-
-            @Override
-            public void onError(String error, Exception exception) {
-                Log.d("DriverHomePage", "Logout error");
-            }
-
-        };
-        return object;
+    //Inflate the menu
+    @Override
+    public boolean onCreateOptionsMenu(Menu menu) {
+        MenuInflater inflator = getMenuInflater();
+        inflator.inflate(R.menu.toolbar_menu, menu);
+        return true;
     }
+
+    @Override
+    public boolean onOptionsItemSelected(MenuItem item) {
+        switch (item.getItemId()) {
+            case R.id.action_settings:
+                startNewActivity(SettingsActivity.class);
+                return true;
+            default:
+                return super.onOptionsItemSelected(item);
+        }
+    }
+
+    private void startNewActivity(Class activity){
+        Intent intent = new Intent(this, activity);
+        startActivity(intent);
+        //finish();
+    }
+
 
 }
