+++ app/src/main/res/layout/activity_login.xml
@@ -11,7 +11,6 @@
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
         android:layout_marginTop="342dp"
-        android:backgroundTint="#317034"
         android:textSize="30sp"
         android:text="@string/sign_in"
         app:layout_constraintBottom_toBottomOf="parent"
