+++ app/src/main/res/layout/activity_trucks.xml
@@ -4,8 +4,9 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
+    android:backgroundTint="@color/primaryColor"
+    android:foregroundTint="#4CAF50"
     tools:context=".TrucksActivity">
-
 
 
     <com.google.android.material.card.MaterialCardView
@@ -14,6 +15,7 @@
         android:layout_height="wrap_content"
         android:layout_margin="8dp"
         tools:ignore="MissingConstraints">
+
 
         <LinearLayout
             android:layout_width="match_parent"
@@ -32,11 +34,10 @@
                     android:id="@+id/name"
                     android:layout_width="wrap_content"
                     android:layout_height="wrap_content"
-                    android:layout_alignParentTop="true"
                     android:layout_alignParentStart="true"
+                    android:layout_alignParentTop="true"
                     android:text="Truck #1"
-                    android:textAppearance="?attr/textAppearanceHeadline6"
-                    />
+                    android:textAppearance="?attr/textAppearanceHeadline6" />
 
                 <ImageView
                     android:id="@+id/imageView3"
@@ -50,10 +51,10 @@
                     android:id="@+id/task"
                     android:layout_width="wrap_content"
                     android:layout_height="wrap_content"
-                    android:layout_alignEnd="@id/imageView3"
                     android:layout_alignBaseline="@id/imageView3"
-                    android:layout_marginBottom="20dp"
+                    android:layout_alignEnd="@id/imageView3"
                     android:layout_marginEnd="45dp"
+                    android:layout_marginBottom="20dp"
                     android:text="Going to grain cart" />
 
                 <TextView
@@ -69,9 +70,9 @@
                     android:id="@+id/imageView"
                     android:layout_width="wrap_content"
                     android:layout_height="wrap_content"
-                    android:layout_toLeftOf="@id/task"
                     android:layout_marginEnd="5dp"
                     android:layout_marginBottom="15dp"
+                    android:layout_toLeftOf="@id/task"
                     app:srcCompat="@drawable/current_task_icon" />
 
                 <ImageView
@@ -121,31 +122,21 @@
 
     </com.google.android.material.card.MaterialCardView>
 
-<!--    <com.google.android.material.floatingactionbutton.FloatingActionButton-->
-<!--        android:id="@+id/fab"-->
-<!--        android:layout_width="wrap_content"-->
-<!--        android:layout_height="wrap_content"-->
-<!--        android:layout_gravity="bottom|end"-->
-<!--        android:layout_margin="16dp"-->
-<!--        app:srcCompat="@drawable/new_icon"/>-->
+    <!--    <com.google.android.material.floatingactionbutton.FloatingActionButton-->
+    <!--        android:id="@+id/fab"-->
+    <!--        android:layout_width="wrap_content"-->
+    <!--        android:layout_height="wrap_content"-->
+    <!--        android:layout_gravity="bottom|end"-->
+    <!--        android:layout_margin="16dp"-->
+    <!--        app:srcCompat="@drawable/new_icon"/>-->
 
-   <com.leinardi.android.speeddial.SpeedDialView
-       android:id="@+id/speedDial"
-       android:layout_width="wrap_content"
-       android:layout_height="wrap_content"
-       android:layout_gravity="bottom|end"
-       app:sdMainFabClosedSrc="@drawable/new_icon"
-       app:sdMainFabOpenedSrc="@drawable/edit_icon"
-       tools:ignore="MissingClass" />
-
-    <Button
-        android:id="@+id/signOutBtn"
+    <com.leinardi.android.speeddial.SpeedDialView
+        android:id="@+id/speedDial"
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
-        android:layout_gravity="bottom|start"
-        android:layout_marginEnd="8dp"
-        android:layout_marginBottom="8dp"
-        android:text="Sign out" />
+        android:layout_gravity="bottom|end"
+        app:sdMainFabClosedSrc="@drawable/new_icon"
+        app:sdMainFabOpenedSrc="@drawable/edit_icon" />
 
 
 </androidx.coordinatorlayout.widget.CoordinatorLayout>
