+++ app/src/main/res/menu/toolbar_menu.xml
@@ -3,7 +3,7 @@
     xmlns:app="http://schemas.android.com/apk/res-auto">
 
     <item
-        android:id="@+id/action_favorite"
+        android:id="@+id/action_settings"
         android:icon="@drawable/ic_settings_icon"
         android:title="Settings"
         app:showAsAction="never"/>
