+++ app/src/main/res/values/colors.xml
@@ -7,4 +7,9 @@
     <color name="teal_700">#FF018786</color>
     <color name="black">#FF000000</color>
     <color name="white">#FFFFFFFF</color>
+
+    <color name="primaryColor">#669e37</color>
+    <color name="primaryLightColor">#97cf65</color>
+    <color name="primaryDarkColor">#366f00</color>
+    <color name="primaryTextColor">#000000</color>
 </resources>
