+++ app/src/main/res/values/strings.xml
@@ -12,4 +12,5 @@
     <string name="TruckN1"></string>
     <string name="mrg_logistics">MRG Logistics</string>
     <string name="sign_in">Sign In</string>
+    <string name="sign_out">Sign Out</string>
 </resources>
