+++ app/src/main/res/values/themes.xml
@@ -2,11 +2,11 @@
     <!-- Base application theme. -->
     <style name="Theme.Team31" parent="Theme.MaterialComponents.DayNight.NoActionBar">
         <!-- Primary brand color. -->
-        <item name="colorPrimary">@color/purple_500</item>
-        <item name="colorPrimaryVariant">@color/purple_700</item>
-        <item name="colorOnPrimary">@color/white</item>
+        <item name="colorPrimary">@color/primaryColor</item>
+        <item name="colorPrimaryVariant">@color/primaryDarkColor</item>
+        <item name="colorOnPrimary">@color/black</item>
         <!-- Secondary brand color. -->
-        <item name="colorSecondary">@color/teal_200</item>
+        <item name="colorSecondary">@color/primaryColor</item>
         <item name="colorSecondaryVariant">@color/teal_700</item>
         <item name="colorOnSecondary">@color/black</item>
         <!-- Status bar color. -->
@@ -14,14 +14,15 @@
         <!-- Customize your theme here. -->
     </style>
 
+
+
+
+
     <style name="Theme.Team31.NoActionBar">
         <item name="windowActionBar">false</item>
         <item name="windowNoTitle">true</item>
     </style>
-
     <style name="Theme.Team31.AppBarOverlay" parent="ThemeOverlay.AppCompat.Dark.ActionBar" />
-
-
     <style name="Theme.Team31.PopupOverlay" parent="ThemeOverlay.AppCompat.Light" />
